Move `main_inner` into datafusion-cli lib for subsequent PR
