+++ datafusion-cli/src/lib.rs
@@ -26,6 +26,7 @@ pub const DATAFUSION_CLI_VERSION: &str = env!("CARGO_PKG_VERSION");
 pub mod catalog;
 pub mod cli_context;
 pub mod command;
+pub mod entry_point;
 pub mod exec;
 pub mod functions;
 pub mod helper;
